backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai>=1.12.0
psycopg2-binary==2.9.9
pymssql==2.2.11
oracledb==2.1.0
python-dotenv==1.0.0
pydantic==2.5.0
httpx>=0.27.0
pandas==2.1.3
openpyxl==3.1.2
python-multipart==0.0.6
requests>=2.31.0

